week2_task/pyproject.toml
[project]
name = "week2"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-json-logger (>=4.0.0,<5.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
